+++ CHANGELOG.md
@@ -5,6 +5,7 @@ SEO42 - Changelog
 
 * Neu: www.seobility.net zu den Tools hinzugefügt
 * Neu: Lang-Presets hinzugefügt: ungarisch, schwedisch und norwegisch, thx@polarpixel
+* Neu: `seo42::getAbsoluteImageFile()` hinzugefügt
 * Neu: FAQ Eintrag: "Warum wird meine Website umgeleitet auf eine andere URL?"
 * Verbessert: No Double Content Redirect: One Domain Only wieder standardmäßig aktiviert und Option hinzugefügt damit dieser Redirect nur noch fürs Frontend greift per default
 * Finetuning